Moved Authorizer into Hooks namespace
